+++ app/src/components/HotKeys/index.ts
@@ -1,6 +1,0 @@
-// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
-//
-// SPDX-License-Identifier: EUPL-1.2
-import HotKeys from './HotKeys';
-
-export default HotKeys;